+++ packages/react-sdk/generator/templates/index.js
@@ -1,1 +1,7 @@
-
+const indexTemplate = `
+export { StoryProvider, useStoryContext } from "./storyProtocolContext";
+<% resources.forEach((resource) => {%>
+export * from "./resources/use<%=resource %>";
+<%})%>
+`;
+module.exports = indexTemplate;
